Add files via upload
The focus of the project revolves around three pivotal aspects:

Overview of the Colleagues: An in-depth look at the existing workforce in the region, encompassing employee details. 

Employee Trends Over Time: An exploration of the dynamic changes in the colleague population over time to uncover insights and trends relevant to Region N07. 

Factors Impacting Future Recruitment: Exploration of the critical factors that can significantly influence future recruitment at the Retail store.